requirements.txt
torch
transformers
pyyaml
psutil
